+++ client/src/components/pages/header-footer/Footer.jsx
@@ -8,7 +8,11 @@ const Footer = () => {
       className="w-100 dark-light mt-5"
     >
       <div className="text-center py-3">
-        © {currentYear} Copyright
+        <p className="text-warning" style={{ fontSize: "0.75rem" }}>
+          Warning: The server and database will be reset{" "}
+          <strong>17. December 19:00 CET </strong>
+        </p>
+        {/* © {currentYear} Copyright */}
         <div className="py-2">
           <a
             target="_blank"