+++ changelog.md
@@ -13,7 +13,7 @@ All notable changes to CyberhackerWarfare4000 will be documented in this file.
 - Earn battery codes auto generates once every day (so you don't lose hours between eacch generation)
 - All ladders includes crypto currency in total wealth
 - Data centers visible in profile
-- Alliance now has a dashboard with ------
+- Alliance dashboard
 - Bodyguards now has "two lifes" before dying
 - Your alive bodyguards can be seen in profile page
 - Small ui changes to notification page